Cliente, TipoCliente, telefones e Enderecos

+++ src/main/java/com/phellipesander/cursomc/entity/Cidade.java
@@ -15,14 +15,14 @@ public class Cidade implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 	
 	@ManyToOne
 	@JoinColumn(name="estado_id")
 	private Estado estado;
 
-	public Cidade(Integer id, String nome, Estado estado) {
+	public Cidade(Long id, String nome, Estado estado) {
 		this.id = id;
 		this.nome = nome;
 		this.estado = estado;
@@ -31,11 +31,11 @@ public class Cidade implements Serializable {
 	public Cidade() {
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
